src/dspygen/agents/coder_agent.py
from dspygen.mixin.fsm.fsm_mixin import FSMMixin, trigger


from enum import Enum, auto


class CoderAgentState(Enum):
    """ Enum representing the states in the coding task lifecycle. """
    ANALYZING_REQUIREMENTS = auto()
    WRITING_CODE = auto()
    TESTING_CODE = auto()
    REFACTORING_CODE = auto()
    HANDLING_ERRORS = auto()
    COMPLETING_TASK = auto()


class CoderAgent(FSMMixin):
    def __init__(self, initial=CoderAgentState.ANALYZING_REQUIREMENTS):
        super().setup_fsm(CoderAgentState, initial=initial)
        self.git_repo = None
        self.code_base = None
        self.tests = None
        self.errors = None
        self.console_output = None
        # Initialize any state-specific attributes if necessary

    @trigger(
        source=CoderAgentState.ANALYZING_REQUIREMENTS,
        dest=CoderAgentState.WRITING_CODE,
        # conditions=['requirements_met'],
        after=['notify_progress']
    )
    def start_coding(self):
        """Transition from analyzing requirements to writing code."""
        print("Starting to write code.")

    @trigger(
        source=CoderAgentState.WRITING_CODE,
        dest=CoderAgentState.TESTING_CODE,
        conditions=['code_complete'],
        after=['run_tests']
    )
    def test_code(self):
        """Transition from writing code to testing code."""
        print("Testing code now.")

    @trigger(
        source=CoderAgentState.TESTING_CODE,
        dest=CoderAgentState.HANDLING_ERRORS,
        conditions=['errors_detected'],
        unless=['all_tests_passed']
    )
    def handle_errors(self):
        """Handle errors if any are detected during testing."""
        print("Handling coding errors.")

    @trigger(
        source=[CoderAgentState.HANDLING_ERRORS, CoderAgentState.TESTING_CODE],
        dest=CoderAgentState.REFACTORING_CODE,
        conditions=['errors_resolved'],
        after=['optimize_code']
    )
    def refactor_code(self):
        """Refactor code after errors are resolved."""
        print("Refactoring code.")

    @trigger(
        source=[CoderAgentState.TESTING_CODE, CoderAgentState.REFACTORING_CODE],
        dest=CoderAgentState.COMPLETING_TASK,
        # conditions=['optimization_complete'],
        after=['finalize_task']
    )
    def complete_task(self):
        """Complete the coding task after successful refactoring."""
        print("Task completed.")

    # Placeholder methods for conditions and actions
    def requirements_met(self):
        return True

    def notify_progress(self):
        print("Notifying progress on the task.")

    def code_complete(self):
        return True

    def run_tests(self):
        print("Running tests on the code.")

    def errors_detected(self):
        return False

    def all_tests_passed(self):
        return True

    def errors_resolved(self):
        return True

    def optimize_code(self):
        print("Optimizing the code.")

    def optimization_complete(self):
        return True

    def finalize_task(self):
        print("Finalizing and wrapping up the task.")


def main():
    # Create an instance of the CoderAgent starting at the initial state
    coder = CoderAgent()

    # Print the initial state
    print(f"Initial State: {coder.state}")

    # Start coding by transitioning from analyzing requirements to writing code
    coder.start_coding()
    print(f"State after starting to code: {coder.state}")

    # Test the code by transitioning from writing code to testing code
    coder.test_code()
    print(f"State after testing code: {coder.state}")

    # Assuming tests pass, move directly to completing task
    # Normally, you would handle errors if they occurred, but this is the happy path
    coder.complete_task()
    print(f"State after completing the task: {coder.state}")

    # Finalize and print confirmation of task completion
    print("Task has been successfully completed!")

# This is a protective measure to ensure the main function is executed only when the script is run directly,
# not when imported as a module.
if __name__ == "__main__":
    main()
